requirements.txt
keyboard==0.13.5
MouseInfo==0.1.3
mss==2.0.0
numpy==1.19.2
opencv-python==4.4.0.44
Pillow==7.2.0
PyAutoGUI==0.9.50
PyAutoIt==0.6.0
PyGetWindow==0.0.8
PyMsgBox==1.0.8
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
PyTweening==1.0.3
